refactor: readSpecialCmsImage
