+++ src/actions/cms/images/create.ts
@@ -6,5 +6,5 @@ import { createCmsImage } from '@/server/cms/images/create'
 
 export async function createCmsImageAction(name: string, image?: Image): Promise<ActionReturn<ExpandedCmsImage>> {
     //TODO: Auth route (very few people should be able to create stand alone cmsImages...)
-    return await createCmsImage(name, image)
+    return await createCmsImage(name, {}, image)
 }
