+++ src/actions/cms/images/read.ts
@@ -1,66 +1,40 @@
 'use server'
-import prisma from '@/prisma'
 import { createActionError, createPrismaActionError } from '@/actions/error'
 import { SpecialCmsImage } from '@prisma/client'
-import type { ExpandedCmsImage } from './Types'
+import type { ExpandedCmsImage } from '@/cms/images/Types'
 import type { ActionReturn } from '@/actions/Types'
+import { readCmsImage, readSpecialCmsImage } from '@/server/cms/images/read'
+import { createCmsImage } from '@/server/cms/images/create'
 
 /**
- * Read a cms image including the image
+ * A action to read a cms image including the image associated with it
  * @param name - name of the cms image the image
  * @returns
  */
-export async function readCmsImage(name: string): Promise<ActionReturn<ExpandedCmsImage>> {
-    //Note this action reates a image link if it does not exist and returns it
-    try {
-        const cmsImage = await prisma.cmsImage.findUnique({
-            where: {
-                name,
-            },
-            include: {
-                image: true,
-            }
-        })
-        if (!cmsImage) return createActionError('NOT FOUND', `${name} Cms Image not found`)
-        return { success: true, data: cmsImage }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function readCmsImageAction(name: string): Promise<ActionReturn<ExpandedCmsImage>> {
+    //TODO: auth on visibilty
+    return await readCmsImage(name)
 }
 
 /**
- * Reads a special cmsImage, if it does not exist it creates it
+ * Action to reads a special cmsImage, if it does not exist it creates it
  * @param special SpecialCmsImage
  * @returns ActionReturn<ExpandedCmsImage>
  */
-export async function readSpecialCmsImage(special: SpecialCmsImage): Promise<ActionReturn<ExpandedCmsImage>> {
+export async function readSpecialCmsImageAction(special: SpecialCmsImage): Promise<ActionReturn<ExpandedCmsImage>> {
     if (!Object.values(SpecialCmsImage).includes(special)) {
         return createActionError('BAD PARAMETERS', `${special} is not special`)
     }
-
-    try {
-        const cmsImage = await prisma.cmsImage.findUnique({
-            where: {
+    const specialRes = await readSpecialCmsImage(special)
+    if (!specialRes.success) {
+        if (specialRes.errorCode === 'NOT FOUND') {
+            return await createCmsImage(special, {
                 special,
-            },
-            include: {
-                image: true,
-            }
-        })
-        if (!cmsImage) {
-            const created = await prisma.cmsImage.create({
-                data: {
-                    name: special,
-                    special,
-                },
-                include: {
-                    image: true
-                }
             })
-            return { success: true, data: created }
         }
-        return { success: true, data: cmsImage }
-    } catch (error) {
-        return createPrismaActionError(error)
+        return specialRes
     }
+    const cmsImage = specialRes.data
+    //TODO: Auth on visibilty
+    return { success: true, data: cmsImage }
 }
